requirements.txt
openai
crewai
langgraph
duckduckgo-search
feedparser
flask
fpdf
python-dotenv
flask
gunicorn
requests
beautifulsoup4
feedparser
groq
